# docs/api-docs.yaml
swagger: "2.0"
info:
  title: BlackSwan-CRUD test API
  version: 1.0.0
schemes:
  - http
host: localhost:8080
basePath: "/api"
paths:
  /users:
    get:
      summary: Returns a list of users.
      produces:
        - application/json
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              type: object
              $ref: '#/definitions/UserResponse'
            description: The found user object array in the database
    post:
      summary: Creates a new user
      parameters:
        - name: User
          description: User to be created
          in: body
          required: true
          schema:
            $ref: "#/definitions/UserRequest"
      produces:
        - application/json
      responses:
        200:
          description: OK, User has been created
          schema:
            $ref: "#/definitions/UserResponse"
            description: The created user object
        409:
          description: User with that username already exists
        422:
          description: Parameters required to create a new user are missing.
  /users/{userid}:
    get:
      summary: Get user's info by userId
      parameters:
        - name: userid
          in: path
          description: ID of the user
          required: true
          type: string
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/UserResponse"
            description: The user object with the given ID
        404:
          description: User not found with the given ID
    put:
      summary: Update existing user info
      parameters:
        - name: User
          in: body
          description: Properties to update on the User
          required: true
          schema:
            $ref: "#/definitions/UserRequest"
        - name: userid
          in: path
          description: ID of the user
          required: true
          type: string
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/UserResponse'
            description: The updated user object
        404:
          description: User not found with the given ID

definitions:
  UserRequest:
    type: object
    properties:
      username:
        type: string
      first_name:
        type: string
      last_name:
        type: string
  UserResponse:
    type: object
    properties:
      username:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      _id:
        type: string